updat main and models

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, String, DateTime, BigInteger
 import datetime
 from database.database import Base
 
@@ -7,7 +7,7 @@ class User(Base):
 
     __tablename__ = "user"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     created_at = Column(DateTime, default=datetime.datetime.utcnow)
     status = Column(String, default="alive")
     status_updated_at = Column(DateTime, default=datetime.datetime.utcnow)
